+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added list of supported boards and generation of quickstart.json file.
 * Added spotflowup.sh and spotflowup.ps1 scripts for workspace setup.
 * Added interfaces for reporting metrics to Spotflow
+* Added device uptime to coredump chunks
 
 ### Removed
 * Removed configdefault for MAIN_STACK_SIZE